task/django_auth/requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
pycparser==2.20
PyJWT==1.5.3
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.8.6
whitenoise==5.2.0
